// @types/graphql-types.d.ts
export interface PageInfo {
  hasNextPage: boolean; /** When paginating, are there more items? */
}
// ...
export interface previous {
  fields?: fields_2 | null; 
  frontmatter?: frontmatter_2 | null; 
}

export interface fields_2 {
  slug?: string | null; 
}

export interface frontmatter_2 {
  title?: string | null; 
}

export interface next {
  fields?: fields_3 | null; 
  frontmatter?: frontmatter_3 | null; 
}
// ...
export interface MarkdownRemark extends Node {
  id: string; /** The id of this node. */
  parent?: Node | null; /** The parent of this node. */
  children?: (Node | null)[] | null; /** The children of this node. */
  internal?: internal_12 | null; 
  frontmatter?: frontmatter_4 | null; 
  rawMarkdownBody?: string | null; 
  fileAbsolutePath?: string | null; 
  fields?: fields_4 | null; 
  html?: string | null; 
  htmlAst?: JSON | null; 
  excerpt?: string | null; 
  headings?: (MarkdownHeading | null)[] | null; 
  timeToRead?: number | null; 
  tableOfContents?: string | null; 
  wordCount?: wordCount | null; 
}
// ...
export interface MarkdownRemarkConnection {
  pageInfo: PageInfo; /** Information to aid in pagination. */
  edges?: (MarkdownRemarkEdge | null)[] | null; /** A list of edges. */
  totalCount?: number | null; 
  distinct?: (string | null)[] | null; 
  group?: (markdownRemarkGroupConnectionConnection | null)[] | null; 
}
/** An edge in a connection. */
export interface MarkdownRemarkEdge {
  node?: MarkdownRemark | null; /** The item at the end of the edge */
  next?: MarkdownRemark | null; /** The next edge in the connection */
  previous?: MarkdownRemark | null; /** The previous edge in the connection */
}
/** A connection to a list of items. */
export interface markdownRemarkGroupConnectionConnection {
  pageInfo: PageInfo; /** Information to aid in pagination. */
  edges?: (markdownRemarkGroupConnectionEdge | null)[] | null; /** A list of edges. */
  field?: string | null; 
  fieldValue?: string | null; 
  totalCount?: number | null; 
}
/** An edge in a connection. */
export interface markdownRemarkGroupConnectionEdge {
  node?: MarkdownRemark | null; /** The item at the end of the edge */
  next?: MarkdownRemark | null; /** The next edge in the connection */
  previous?: MarkdownRemark | null; /** The previous edge in the connection */
}